swiftkyc/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

redis==5.2.0

pydantic==2.9.2
pydantic-settings==2.6.1

python-multipart==0.0.17  


rq==1.15.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.3.0

easyocr==1.7.2

